+++ cmd/hidden_lake/applications/filesharer/internal/stream/errors.go
@@ -19,4 +19,5 @@ var (
 	ErrRetryFailed      = &SStreamError{"retry failed"}
 	ErrInvalidWhence    = &SStreamError{"invalid whence"}
 	ErrNegativePosition = &SStreamError{"negative position"}
+	ErrGetMessageLimit  = &SStreamError{"get message limit"}
 )