+++ cmd/hidden_lake/applications/filesharer/internal/stream/types.go
@@ -3,3 +3,9 @@ package stream
 import "io"
 
 type IReadSeeker io.ReadSeeker
+
+type IFileInfo interface {
+	GetName() string
+	GetHash() string
+	GetSize() uint64
+}